Read UI constants from a plist, move version string generation into a tiny service.
