+++ TriviaQuiz/Config/UIConfigData.plist
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>foo</key>
+	<string>bar</string>
+	<key>LayerSize</key>
+	<dict>
+		<key>cornerRadius</key>
+		<integer>15</integer>
+		<key>borderWidth</key>
+		<integer>2</integer>
+	</dict>
+</dict>
+</plist>
